services/camera-adapter/requirements.txt
redis==5.0.1
requests==2.31.0
xmltodict==0.13.0
minio==7.2.0
python-dotenv==1.0.0
tenacity==8.3.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-logging